Match fcmla with more generic contract pattern
Use an approach similar to how reassoc is handled.
However, in this case, we need to maintain the structure of the operations
so instead of collecting a set of multiplications to be added together,
we build a stack of multiplications that will be added in the stack order.

Compared to the old approach, the depth of the stack can be 1
(to match unpaired single partial multiplication) and can also be
arbitrarily deep (to match longer complex computations).
Similar to the reassoc case, we can also walk the stack to find
complex pairs of common terms that may be more than one level
away from each other.