Improve negative addend handling for reassoc case
If we couldn't find a positive addend, we could simply find a negative one
and use that as the accumulator.
In the worst case we may need to add a negation to the final result
but we'll get rid of an add/sub between addends and a zero initialization
of the accumulator.